Initial Dominos API POC (#1)
I had to manually copy [pizzapi](https://pypi.org/project/pizzapi)
because there was an unfixed bug, and copying was the suggested solution
in the repo. So far, all we're doing is creating a customer and address,
then finding the nearest store and printing the menu. We're not going to
move forward with the Dominos side of the project until confirming what
user interactions we want to support.

From what I can tell so far, doing items like building your own pizza
will be difficult and we might be limited to pre-built options.