LCD Functions (#5)
I wrote this code (@hardins01), but the commits are made by @loganrower
because I made the commits on his Raspberry Pi.

I created an LCD class to make it very simple to interact with the
display. The class also manages the logic of automatic scrolling of
messages if they're too long for the width of the display. I added
documentation for the LCD class in `hardware/lcd/lcd.md`.

---------

Co-authored-by: loganrower <lkrower2000@gmail.com>